Attempt another round of dom update file
